requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
annotated-types==0.7.0
    # via
    #   -r requirements.in
    #   pydantic
anyio==4.9.0
    # via
    #   -r requirements.in
    #   httpx
    #   starlette
base58==2.1.1
    # via -r requirements.in
certifi==2025.6.15
    # via
    #   -r requirements.in
    #   httpcore
    #   httpx
click==8.2.1
    # via
    #   -r requirements.in
    #   prisma
    #   uvicorn
colorama==0.4.6
    # via -r requirements.in
dnspython==2.7.0
    # via
    #   -r requirements.in
    #   email-validator
email-validator==2.2.0
    # via -r requirements.in
fastapi==0.115.13
    # via -r requirements.in
h11==0.16.0
    # via
    #   -r requirements.in
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via
    #   -r requirements.in
    #   httpx
httptools==0.6.4
    # via -r requirements.in
httpx==0.28.1
    # via
    #   -r requirements.in
    #   prisma
idna==3.10
    # via
    #   -r requirements.in
    #   anyio
    #   email-validator
    #   httpx
itsdangerous==2.2.0
    # via -r requirements.in
jinja2==3.1.6
    # via
    #   -r requirements.in
    #   prisma
json5==0.12.0
    # via -r requirements.in
markupsafe==3.0.2
    # via
    #   -r requirements.in
    #   jinja2
nodeenv==1.9.1
    # via prisma
prisma==0.15.0
    # via -r requirements.in
pydantic==2.11.7
    # via
    #   -r requirements.in
    #   fastapi
    #   prisma
    #   pydantic-settings
pydantic-core==2.33.2
    # via
    #   -r requirements.in
    #   pydantic
pydantic-settings==2.10.0
    # via -r requirements.in
python-dotenv==1.1.0
    # via
    #   -r requirements.in
    #   prisma
    #   pydantic-settings
python-multipart==0.0.20
    # via -r requirements.in
pyyaml==6.0.2
    # via -r requirements.in
six==1.17.0
    # via -r requirements.in
sniffio==1.3.1
    # via
    #   -r requirements.in
    #   anyio
starlette==0.46.2
    # via
    #   -r requirements.in
    #   fastapi
tomlkit==0.13.3
    # via prisma
typing-extensions==4.14.0
    # via
    #   -r requirements.in
    #   anyio
    #   fastapi
    #   prisma
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.1
    # via
    #   -r requirements.in
    #   pydantic
    #   pydantic-settings
uvicorn==0.34.3
    # via -r requirements.in
uvloop==0.21.0
    # via -r requirements.in
websockets==15.0.1
    # via -r requirements.in
